Move tests for removing tags in favor of digests
The code is now in utils module, move the test to the corresponding
part. A new test case is added to ensure the stripping is called when
extracting labels.
