+++ tests/test_containers.py
@@ -9,19 +9,7 @@ from unittest import mock
 
 import pytest
 
-from rpm_lockfile import containers, utils
-
-
-@pytest.mark.parametrize(
-    "image_spec,expected",
-    [
-        ("example.com/image:latest", "example.com/image:latest"),
-        ("example.com/image@sha256:abcdef", "example.com/image@sha256:abcdef"),
-        ("example.com/image:latest@sha256:0123456", "example.com/image@sha256:0123456"),
-    ],
-)
-def test_strip_tag(image_spec, expected):
-    assert utils.strip_tag(image_spec) == expected
+from rpm_lockfile import containers
 
 
 @pytest.fixture
